+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-### 4.8.0(75)
+### [4.8.0(75)](https://github.com/VakhoKontridze/VCore/releases/tag/4.8.0) — *2023 04 13*
 
 General